Fixed var statement typo
The opening `var` statement declared a `countAsync` variable which is not used anywhere else in the file. But there is an `isAsync` variable set to false a few lines down, which isn't declared using `var`. I assume this is a typo (since this code fails under "use strict" with a ReferenceError), and I renamed the variable declared in `var` to `isAsync`.